+++ projects/components/src/lib/action-dropdown/action-dropdown-option.component.ts
@@ -1,7 +1,7 @@
 import { Component, ElementRef, EventEmitter, Input, Output, ViewChild } from '@angular/core';
 
 @Component({
-  selector: 'mas-dropdown-option',
+  selector: 'mas-action-dropdown-option',
   template: `
     <img src="{{ imageUrl }}" class="mas-dropdown-option-image" *ngIf="type === 'image'" />
     <i class="{{ icon }} mas-icon mas-dropdown-option_icon" *ngIf="type === 'icon' && icon !== ''"></i>
@@ -10,7 +10,6 @@ import { Component, ElementRef, EventEmitter, Input, Output, ViewChild } from '@
     ></span>
   `,
   host: {
-    '[class.mas-dropdown-option--selected]': '_selected',
     '[class.mas-dropdown-option--medium]': 'size === "M"',
     '[class.mas-dropdown-option--small]': 'size === "S"',
     '[class.mas-dropdown-option--focus]': 'isFocus',
@@ -20,6 +19,27 @@ import { Component, ElementRef, EventEmitter, Input, Output, ViewChild } from '@
   },
 })
 export class MasActionDropdownOption {
+  /**
+   * Image for MasActionDropdownOption - type: image.
+   */
+  @Input() imageUrl: string = '';
+  /**
+   * Icon for MasActionDropdownOption - type: icon.
+   */
+  @Input() icon: string = '';
+  /**
+   * MasActionDropdown's type include: 'checkbox' | 'image' | 'icon' | 'text'.
+   */
+  @Input() type: 'checkbox' | 'image' | 'icon' | 'text' = 'text';
+  /**
+   * Whether the component is disabled.
+   */
+  @Input() disabled: boolean;
+  /**
+   * Callback to invoke when value of MasActionDropdownOption clicked.
+   */
+  @Output() onClick: EventEmitter<MasActionDropdownOptionClickEvent> = new EventEmitter();
+  @ViewChild('text', { static: true }) _text: ElementRef<HTMLElement> | undefined;
   public size: 'S' | 'M' = 'S';
   _isFocus: boolean = false;
   get isFocus() {
@@ -28,20 +48,16 @@ export class MasActionDropdownOption {
   set isFocus(value: boolean) {
     this._isFocus = value;
   }
-  @Input() imageUrl: string = '';
-  @Input() icon: string = '';
-  @Input() type: 'checkbox' | 'image' | 'icon' | 'text' = 'text';
-  @Input() disabled: boolean;
-  @Input() onChange = () => {};
-  @Output() change: EventEmitter<any> = new EventEmitter();
-  @Output() onClick: EventEmitter<MouseEvent> = new EventEmitter<MouseEvent>();
-  @ViewChild('text', { static: true }) _text: ElementRef<HTMLElement> | undefined;
   constructor(public elRef: ElementRef) {}
   get text(): any {
     return this._text?.nativeElement.outerText;
   }
   onOptionClick(event: Event) {
-    this.onChange();
-    this.change.emit({ originalEvent: event, option: this });
+    this.onClick.emit({ originalEvent: event, option: this });
   }
+}
+
+export class MasActionDropdownOptionClickEvent {
+  originalEvent: Event;
+  option?: MasActionDropdownOption
 }