+++ projects/components/src/lib/action-dropdown/action-dropdown.component.ts
@@ -6,7 +6,6 @@ import {
   EventEmitter,
   HostListener,
   Input,
-  OnInit,
   Output,
   QueryList,
 } from '@angular/core';
@@ -33,27 +32,44 @@ import { Observable, Subject, defer, merge, startWith, switchMap, take, takeUnti
     role: 'select',
   },
 })
-export class MasActionDropdown implements OnInit, AfterContentInit {
+export class MasActionDropdown implements AfterContentInit {
+  /**
+   * Leading icon for MasActionDropdown.
+   */
   @Input() leadingIcon: string = '';
+  /**
+   * Label for MasActionDropdown.
+   */
   @Input() labelText: string = '';
+  /**
+   * A unique id for the MasActionDropdown. If none is supplied, it will be auto-generated.
+   */
   @Input() identifier = `action-dropdown-${MasActionDropdown.dropdownCount++}`;
+  /**
+   * Whether the component is disabled.
+   */
   @Input() disabled: boolean = false;
+  /**
+   * The size of the action-dropdown. Availabel options: 'S', 'M'. Default: 'S'.
+   */
   @Input() size: 'M' | 'S' = 'S';
-
-  @Output() change = new EventEmitter();
-  readonly _destroy = new Subject<void>();
+  /**
+   * Callback to invoke when value of MasInputDropdown changes.
+   */
+  @Output() onChange: EventEmitter<MasActionDropdownChangeEvent> = new EventEmitter();
   @ContentChildren(MasActionDropdownOption) options: QueryList<MasActionDropdownOption>;
   static dropdownCount: number = 0;
   protected _panelOpen: boolean = false;
+  protected focusIndex: number = -1;
+  readonly _destroy = new Subject<void>();
   private _ngZone: any;
-  focusIndex: number = -1;
 
   constructor(protected eRef: ElementRef) {}
   ngAfterContentInit(): void {
     this.options.changes.pipe(startWith(null), takeUntil(this._destroy)).subscribe(() => {
       const changeOrDestroyed = merge(this.options.changes, this._destroy);
       this.optionChanges.pipe(takeUntil(changeOrDestroyed)).subscribe((event: any) => {
-        this._onSelect(event.option);
+        this._onChange(event?.originalEvent, event?.option)
         if (this._panelOpen) {
           this.close();
           this.focusOut();
@@ -64,13 +80,12 @@ export class MasActionDropdown implements OnInit, AfterContentInit {
     this.options.map((option) => (option.size = this.size));
   }
 
-  ngOnInit() {}
   readonly optionChanges: Observable<any> = defer(() => {
     const options = this.options;
     if (options) {
       return options.changes.pipe(
         startWith(options),
-        switchMap(() => merge(...options.map((option) => option.change)))
+        switchMap(() => merge(...options.map((option) => option.onClick)))
       );
     }
     return this._ngZone.onStable.pipe(
@@ -79,12 +94,8 @@ export class MasActionDropdown implements OnInit, AfterContentInit {
     );
   }) as Observable<any>;
 
-  // todo: declare function
-  onChange(event: any) {
-    this.change.emit(event);
-  }
-  private _onSelect(option: MasActionDropdownOption): void {
-    this.onChange(option);
+  _onChange(event: Event, option: MasActionDropdownOption) {
+    this.onChange.emit({originalEvent: event, option});
   }
   _onToggle() {
     this.focusIndex = -1;
@@ -128,7 +139,7 @@ export class MasActionDropdown implements OnInit, AfterContentInit {
         // choose that action
         event.preventDefault();
         if (this.options.get(this.focusIndex) !== undefined) {
-          this._onSelect(this.options.get(this.focusIndex)!);
+          this._onChange(event, this.options.get(this.focusIndex)!);
           this.options.get(this.focusIndex)!.isFocus = false;
           this._onToggle();
         } else {
@@ -141,4 +152,8 @@ export class MasActionDropdown implements OnInit, AfterContentInit {
         break;
     }
   }
+}
+export class MasActionDropdownChangeEvent {
+  originalEvent: Event;
+  option?: MasActionDropdownOption 
 }